Fixed permissions on the .py files. Removed pgoapi, added missing utility functions to utils.
