+++ Easy Setup/requirements.txt
@@ -8,11 +8,9 @@ itsdangerous==0.24
 peewee==2.8.1
 wsgiref==0.1.2
 geopy==1.11.0
-protobuf==2.6.1
-requests==2.10.0
 s2sphere==0.2.4
 gpsoauth==0.3.0
-protobuf-to-dict==0.1.0
 PyMySQL==0.7.5
 flask-cors==2.1.2
 flask-compress==1.3.0
+git://github.com/tejado/pgoapi.git@v1.1.0#egg=pgoapi
