+++ pogom/search.py
@@ -9,14 +9,16 @@ from threading import Thread, Lock
 from queue import Queue
 
 from pgoapi import PGoApi
-from pgoapi.utilities import f2i, get_cellid
+from pgoapi.utilities import f2i
+
+from pogom.utils import get_cellid
 
 from . import config
 from .models import parse_map
 
 log = logging.getLogger(__name__)
 
-TIMESTAMP = '\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000'
+#TIMESTAMP = '\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000'
 api = PGoApi()
 
 #Constants for Hex Grid
@@ -39,7 +41,7 @@ def send_map_request(api, position):
         api.set_position(*position)
         api.get_map_objects(latitude=f2i(position[0]),
                             longitude=f2i(position[1]),
-                            since_timestamp_ms=TIMESTAMP,
+                            since_timestamp_ms=[0,] * len(get_cellid(position[0], position[1])),
                             cell_id=get_cellid(position[0], position[1]))
         return api.call()
     except Exception as e:
