+++ pogom/utils.py
@@ -13,6 +13,10 @@ import platform
 import logging
 import shutil
 
+from importlib import import_module
+from s2sphere import CellId, LatLng
+from geopy.geocoders import GoogleV3
+
 from . import config
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s [%(module)11s] [%(levelname)7s] %(message)s')
@@ -139,3 +143,36 @@ def get_pokemon_name(pokemon_id):
             get_pokemon_name.names = json.loads(f.read())
 
     return get_pokemon_name.names[str(pokemon_id)]
+
+def get_pos_by_name(location_name):
+    prog = re.compile("^(\-?\d+\.\d+)?,\s*(\-?\d+\.\d+?)$")
+    res = prog.match(location_name)
+    latitude, longitude, altitude = None, None, None
+    if res:
+        latitude, longitude, altitude = float(res.group(1)), float(res.group(2)), 0
+    elif location_name:
+        geolocator = GoogleV3()
+        loc = geolocator.geocode(location_name)
+        if loc:
+            latitude, longitude, altitude = loc.latitude, loc.longitude, loc.altitude
+
+    return (latitude, longitude, altitude)
+
+def get_class(cls):
+    module_, class_ = cls.rsplit('.', 1)
+    class_ = getattr(import_module(module_), class_)
+    return class_
+
+def get_cellid(lat, long):
+    origin = CellId.from_lat_lng(LatLng.from_degrees(lat, long)).parent(15)
+    walk = [origin.id()]
+
+    # 10 before and 10 after
+    next = origin.next()
+    prev = origin.prev()
+    for i in range(10):
+        walk.append(prev.id())
+        walk.append(next.id())
+        next = next.next()
+        prev = prev.prev()
+    return sorted(walk)
